refactor : ApiResponseDto 위치 이동
